so: 32-bit support

+++ so/math/bits/bits.go
@@ -10,7 +10,7 @@
 // [math/bits]: https://github.com/golang/go/blob/go1.26.1/src/math/bits/bits.go
 package bits
 
-const uintSize = 32 << (^uint(0) >> 63) // 32 or 64
+const uintSize = 32 << (uint64(^uint(0)) >> 63) // 32 or 64
 
 // UintSize is the size of a uint in bits.
 const UintSize = uintSize
